Move code from CreateCapsulePage to CreateCapsule component. Rename CreateCapsulePage to CapsulesPage.
