+++ frontend/src/pages/CapsulesPage.jsx
@@ -1,0 +1,4 @@
+/**
+ * 
+ * Capsules page
+ */
